Added click detection stuff
New functions can detect mouse clicks, transform them to board coordinates, and determine what square was clicked from there. Works fairly well, returns Z0 if the click was outside of the range of the board. Temporarily added in something to just print these squares, but we can use the output of coords_to_square to feed into the chess engine. I will also likely add a success bool to that function to determine whether it was a legitimate square.